refactor: change resultlsit tests
